backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.2.0

# STP to GLB conversion dependencies
cadquery>=2.4.0
pygltflib>=1.16.0
numpy>=1.24.0
trimesh>=3.20.0

# DWG conversion dependencies (optional)
ezdxf>=1.0.0

